Use a PUT request to update settings rather than socket.io
Since ScribTeX was only using the socket connection for settings and this was
otherwise causing some trouble, migrate over to using a simple AJAX PUT request
to update the settings and do away with needing a socket connection all together.
Something else now needs to call all the initialization parts because it is no longer
triggered in ext/connect/connect.js. For example:
    ide.onLine = true;
    ide.inited = true;
    ide.dispatchEvent("beforeonline");
    ide.dispatchEvent("afteronline");
    stServerConnected.activate();

Conflicts:

	client/core/ide.js
	client/ext/settings/settings.js